+++ cli/Sources/TuistEnvKey/EnvKey.swift
@@ -130,6 +130,7 @@ public enum EnvKey: String, CaseIterable {
     case inspectTestPath = "TUIST_INSPECT_TEST_PATH"
     case inspectTestDerivedDataPath = "TUIST_INSPECT_TEST_DERIVED_DATA_PATH"
     case inspectTestResultBundlePath = "TUIST_INSPECT_TEST_RESULT_BUNDLE_PATH"
+    case inspectTestMode = "TUIST_INSPECT_TEST_MODE"
 
     // RUN