+++ cli/Sources/TuistKit/Services/TestService.swift
@@ -17,6 +17,7 @@ import TuistServer
 import TuistSupport
 import TuistXCResultService
 import XcodeGraph
+import XCResultParser
 
 import struct TSCUtility.Version
 
@@ -295,7 +296,8 @@ public struct TestService { // swiftlint:disable:this type_body_length
             ignoreBinaryCache: ignoreBinaryCache,
             ignoreSelectiveTesting: ignoreSelectiveTesting,
             cacheStorage: cacheStorage,
-            destination: destination
+            destination: destination,
+            schemeName: schemeName
         )
 
         Logger.current.notice("Generating project for testing", metadata: .section)
@@ -1499,6 +1501,7 @@ public struct TestService { // swiftlint:disable:this type_body_length
         let test = try await createTestService.createTest(
             fullHandle: fullHandle,
             serverURL: serverURL,
+            id: nil,
             testSummary: testSummary,
             buildRunId: nil,
             gitBranch: gitInfo.branch,