+++ server/lib/tuist/environment.ex
@@ -562,6 +562,14 @@ defmodule Tuist.Environment do
     get([:processor, :webhook_secret], secrets)
   end
 
+  def xcode_processor_url(secrets \\ secrets()) do
+    get([:xcode_processor, :url], secrets)
+  end
+
+  def xcode_processor_webhook_secret(secrets \\ secrets()) do
+    get([:xcode_processor, :webhook_secret], secrets)
+  end
+
   def clickhouse_flush_interval_ms(secrets \\ secrets()) do
     case get([:clickhouse, :flush_interval_ms], secrets) do
       flush_interval when is_binary(flush_interval) -> String.to_integer(flush_interval)